+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## main
 
+### Bug fix
+
+- Print variants properly in debug SDL.
+
 ## 0.4.0
 
 ### New Features