+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features
 
 - [LSP] Add autocompletion for `ctx: ResGraphContext.context` to make using context easier.
+- [LSP] Context aware snippets for inserting new query fields and full mutations.
 
 ### Bug fix